+++ documentation/source/CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.12.0] - 2021-07-06
+
+### Added
+
+- `why` key for message overrides.
+- Disk storage of sessions storage across runners.  See docs for more info.
+
 ## [0.11.0] - 2021-07-03
 
 ### Added